agents/auto_code_agent/requirements.txt
openai==0.28
python-dotenv
